+++ src/treecf/__init__.py
@@ -31,7 +31,7 @@ from treecf.plausibility import Plausibility
 from treecf.regions import RecourseRegion
 from treecf.targets import Target
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 __all__ = [
     "AllowMissing",